requirements.txt
pandas==2.3.0
bs4==0.0.2
requests==2.32.4
jpholiday==1.0.2
lxml==5.4.0
python-dotenv==1.1.0